+++ scrapper/src/main/java/edu/java/scrapper/client/GithubClient.java
@@ -1,6 +1,6 @@
 package edu.java.scrapper.client;
 
-import edu.java.scrapper.model.github.RepositoryResponse;
+import edu.java.scrapper.dto.github.RepositoryResponse;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.service.annotation.GetExchange;
