+++ scrapper/src/main/java/edu/java/scrapper/dto/bot/RemoveLinkRequest.java
@@ -1,4 +1,4 @@
-package edu.java.scrapper.dto.request;
+package edu.java.scrapper.dto.bot;
 
 import java.net.URI;
