+++ scrapper/src/main/java/edu/java/scrapper/dto/github/RepositoryResponse.java
@@ -1,4 +1,4 @@
-package edu.java.scrapper.model.github;
+package edu.java.scrapper.dto.github;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
