+++ scrapper/src/main/java/edu/java/scrapper/dto/stackoverflow/QuestionResponse.java
@@ -1,4 +1,4 @@
-package edu.java.scrapper.model.stackoverflow;
+package edu.java.scrapper.dto.stackoverflow;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
